test: Add with_local_registry to separate registry in test_query.
