+++ graphene_mongo/tests/utils.py
@@ -1,0 +1,17 @@
+from .. import registry
+
+
+def with_local_registry(func):
+    def inner(*args, **kwargs):
+        old = registry.get_global_registry()
+        registry.reset_global_registry()
+        try:
+            retval = func(*args, **kwargs)
+        except Exception as e:
+            registry.registry = old
+            raise e
+        else:
+            registry.registry = old
+            return retval
+    return inner
+
